+++ src/pages/Index.vue
@@ -1,8 +1,5 @@
 <template>
 <Layout>
-  <div>
-    <h1>Home Page</h1>
-  </div>
   <!-- <a href="https://gridsome.org/docs/" target="_blank" rel="noopener">Gridsome Docs</a> -->
 </Layout>
 </template>